+++ common/config/conf_min_shared.go
@@ -60,10 +60,6 @@ func NewDefaultMinimumRepoConfig() MinimumRepoConfig {
 					RepoPath: "./ipfs",
 				},
 			},
-			Redis: RedisConfig{
-				Enabled: false,
-				Shards:  []RedisShardConfig{},
-			},
 		},
 		AccessTokens: AccessTokenConfig{
 			MaxCacheTimeSeconds: 0,