+++ internal_cache/instance.go
@@ -1,7 +1,6 @@
 package internal_cache
 
 import (
-	"github.com/getsentry/sentry-go"
 	"sync"
 
 	"github.com/sirupsen/logrus"
@@ -20,25 +19,9 @@ func Get() ContentCache {
 		if config.Get().Redis.Enabled {
 			logrus.Info("Setting up Redis cache")
 			instance = NewRedisCache(config.Get().Redis)
-		} else if config.Get().Features.Redis.Enabled {
-			logrus.Info("Setting up Redis cache")
-
-			warnMsg := "Your configuration uses a legacy approach for enabling Redis support. Please move this to the root config or visit #media-repo:t2bot.io for assistance."
-			logrus.Warn(warnMsg)
-			sentry.CaptureMessage(warnMsg)
-
-			instance = NewRedisCache(config.Get().Features.Redis)
-		} else if !config.Get().Downloads.Cache.Enabled {
+		} else {
 			logrus.Warn("Cache is disabled - setting up a dummy instance")
 			instance = NewNoopCache()
-		} else {
-			logrus.Info("Setting up in-memory cache")
-
-			warnMsg := "The built-in cache mechanism is being removed in a future version. Please set up Redis as a cache mechanism."
-			logrus.Warn(warnMsg)
-			sentry.CaptureMessage(warnMsg)
-
-			instance = NewMemoryCache()
 		}
 	})